en uuuuhuere chnorz

+++ nc_benchmark.sh
@@ -69,10 +69,14 @@ CURL="curl -k -s -u$USR:$PW"
 UL_BLOCK_ASSEMBLING_MAX_WAIT=60
 
 # prepare local benchmark dirs
-test -d "$LOCAL_DIR" && rm -rf "$LOCAL_DIR"
-mkdir -p "$LOCAL_DIR/small_files"
-dd if=/dev/urandom of="$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb" bs=1M count=$TEST_BLOCK_SIZE_MB >/dev/null 2>&1
-md5sum "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb" > "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.md5sum"
+test -d "$LOCAL_DIR/small_files" && rm -rf "$LOCAL_DIR/small_files" && mkdir -p "$LOCAL_DIR/small_files"
+
+touch "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb"
+if [ $(( $(stat --printf="%s" "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb" ) / 1024 / 1024 +1 )) -lt $TEST_BLOCK_SIZE_MB ]
+then
+   dd if=/dev/urandom of="$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb" bs=1M count=$TEST_BLOCK_SIZE_MB >/dev/null 2>&1
+fi
+test -f "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.md5sum" || md5sum "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb" > "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.md5sum"
 ls -l $LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb > "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.ls"
 for i in $(seq 1 $TEST_FILES_COUNT)
 do
@@ -115,27 +119,27 @@ echo "$D;$BURL;$USR;UPLOAD;Assembling time $TEST_BLOCK_SIZE_MB.mb;;$UL_BLOCK_ASS
 
 # run block download test
 echo download $TEST_BLOCK_SIZE_MB MB starting: $(date '+%Y.%m.%d %H:%M:%S')
-DL_BLOCK_SPEED=$($CURL --limit-rate $SPEED_LIMIT_DOWN -w '%{speed_download}' "$DAV_REMOTE_BENCH_DIR/$TEST_BLOCK_SIZE_MB.mb" -o "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb" | cut -d. -f1)
+DL_BLOCK_SPEED=$($CURL --limit-rate $SPEED_LIMIT_DOWN -w '%{speed_download}' "$DAV_REMOTE_BENCH_DIR/$TEST_BLOCK_SIZE_MB.mb" -o "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.download" | cut -d. -f1)
 DL_BLOCK_SPEED=$(( $DL_BLOCK_SPEED / 1024 )) # kbyte per sec
 D="$(date '+%Y.%m.%d %H:%M:%S')"
 echo "$D;$BURL;$USR;DOWNLOAD;Block $TEST_BLOCK_SIZE_MB MB;;$DL_BLOCK_SPEED KByte/s;$SPEED_LIMIT_DOWN" >>  $LOCAL_LOG_FILE
 echo download $TEST_BLOCK_SIZE_MB MB finished: $(date '+%Y.%m.%d %H:%M:%S')
-md5sum "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb" > "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.md5sum.after"
-ls -l $LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb > "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.ls.after"
+md5sum "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.download" > "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.md5sum.download"
+ls -l $LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.download > "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.ls.download"
 echo "------ DETAILS BEFORE UPLOAD BIG FILE ------"
 cat $LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.md5sum
 cat $LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.ls
 echo "------ DETAILS AFTER DOWNLOAD BIG FILE ------"
-cat $LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.md5sum.after
-cat $LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.ls.after
+cat $LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.md5sum.download
+cat $LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.ls.download
 
-md5sum --check "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.md5sum" 2>&1 >/dev/null
+if [ $(cat $LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.md5sum.download | cut -d\  -f1) != $(cat $LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.md5sum | cut -d\  -f1) ]
 if [ $? -ne 0 ]
 then 
    DL_BLOCK_SPEED="md5sum error"
 fi
-rm -f "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb"
-rm -f "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.md5sum"
+rm -f "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.download"
+rm -f "$LOCAL_DIR/$TEST_BLOCK_SIZE_MB.mb.md5sum.download"
 
 # run small file upload test
 UL_ERROR_CNT=0
